+++ src/main/java/net/earlystage/mixin/SheepEntityMixin.java
@@ -1,0 +1,46 @@
+package net.earlystage.mixin;
+
+import net.earlystage.init.ItemInit;
+import net.minecraft.entity.EntityType;
+import net.minecraft.entity.passive.AnimalEntity;
+import net.minecraft.entity.passive.SheepEntity;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.item.ItemStack;
+import net.minecraft.sound.SoundCategory;
+import net.minecraft.util.ActionResult;
+import net.minecraft.util.Hand;
+import net.minecraft.world.World;
+import net.minecraft.world.event.GameEvent;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+@Mixin(SheepEntity.class)
+public abstract class SheepEntityMixin extends AnimalEntity {
+
+    public SheepEntityMixin(EntityType<? extends AnimalEntity> entityType, World world) {
+        super(entityType, world);
+    }
+
+    @Inject(method = "interactMob", at = @At("HEAD"), cancellable = true)
+    private void interactMobMixin(PlayerEntity player, Hand hand, CallbackInfoReturnable<ActionResult> info) {
+        ItemStack itemStack = player.getStackInHand(hand);
+        if (itemStack.isOf(ItemInit.STONE_SHEARS)) {
+            if (!this.getWorld().isClient() && this.isShearable()) {
+                this.sheared(SoundCategory.PLAYERS);
+                this.emitGameEvent(GameEvent.SHEAR, player);
+                itemStack.damage(1, player, SheepEntity.getSlotForHand(hand));
+                info.setReturnValue(ActionResult.SUCCESS);
+            }
+        }
+    }
+
+    @Shadow
+    public abstract void sheared(SoundCategory shearedSoundCategory);
+
+    @Shadow
+    public abstract boolean isShearable();
+
+}